💥拆分OperatorResourcesHelper，返回Dictionary的方法改为返回ImmutableDictionary 🔖0.2.0.0
