+++ src/ArknightsResources.Utility.csproj
@@ -16,9 +16,9 @@
     <PackageIcon>icon.png</PackageIcon>
     <EmbedUntrackedSources>true</EmbedUntrackedSources>
     <IsTrimmable>true</IsTrimmable>
-    <Version>0.1.12.0-alpha</Version>
-    <AssemblyVersion>0.1.12.0</AssemblyVersion>
-    <FileVersion>0.1.12.0</FileVersion>
+    <Version>0.2.0.0-alpha</Version>
+    <AssemblyVersion>0.1.0.0</AssemblyVersion>
+    <FileVersion>0.2.0.0</FileVersion>
     <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
     <IncludeSymbols>True</IncludeSymbols>
     <SymbolPackageFormat>snupkg</SymbolPackageFormat>
@@ -37,7 +37,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="ArknightsResources.CustomResourceHelpers" Version="0.1.10-alpha" />
+    <PackageReference Include="ArknightsResources.CustomResourceHelpers" Version="0.2.0-alpha" />
     <PackageReference Include="Fmod5Sharp" Version="3.0.1" />
     <PackageReference Include="K4os.Compression.LZ4" Version="1.3.5" />
     <PackageReference Include="SixLabors.ImageSharp" Version="2.1.3" />
